requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==4.0.4
django-allauth==0.51.0
django-cors-headers==3.12.0
django-extensions==3.1.5
django-rest-authtoken==2.1.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docker==5.0.3
executing==0.8.3
gunicorn==20.1.0
idna==3.3
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
mysqlclient==2.1.0
oauthlib==3.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
PyJWT==1.7.1
PyMySQL==1.0.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
traitlets==5.2.2.post1
urllib3==1.26.9
wcwidth==0.2.5
websocket-client==1.3.3